Placing and exporting a repeating function in a separate file
