+++ src/components/Clusters/components/AddClusterWizard.js
@@ -29,6 +29,7 @@ import { ClusterPreview } from './ClusterPreview';
 
 import './AddClusterWizard.scss';
 import { isFormOpenState } from 'state/formOpenAtom';
+import { checkAuthRequiredInputs } from '../helper';
 
 export function AddClusterWizard({
   kubeconfig,
@@ -158,19 +159,7 @@ export function AddClusterWizard({
   const checkRequiredInputs = () => {
     // setTimeout is used to delay and ensure that the form validation runs after the state updates.
     setTimeout(() => {
-      const invalidList = authFormRef?.current?.querySelectorAll(':invalid');
-      const scopes = authFormRef?.current?.querySelector(
-        '[accessible-name="Scopes"]',
-      );
-      const scopesValid = [
-        ...(scopes?.querySelectorAll('ui5-input') ?? []),
-      ]?.filter(el => el?.value);
-      const isScopesInvalid = scopes && !scopesValid?.length;
-      if (invalidList?.length || isScopesInvalid) {
-        setHasInvalidInputs(true);
-      } else {
-        setHasInvalidInputs(false);
-      }
+      checkAuthRequiredInputs(authFormRef, setHasInvalidInputs);
     });
   };
